CAMEL-11930: camel-spring and camel-blueprint: ...
...make `test` goal execute successfully

Ant task to manipulate schemas fails if the `schema1.xsd` doesn't exist
and this could be because it was moved by previous execution. That is
running `mvn test` after `mvn install` fails because `schema1.xsd` no
longer exists.
This adds a condition so if `schema1.xsd` is not present schema
manipulation is skipped.
Affected components are `camel-spring` and `camel-blueprint`.